Extracts inserted video if detected (generates boxvideo.mp4 and outervideo.mp4)

+++ python/experiment.py
@@ -27,6 +27,8 @@ allimages = []
 first = True
 total_image = 0
 
+debug = False
+
 while(cap.isOpened()):
 
     counter += 1
@@ -45,16 +47,6 @@ while(cap.isOpened()):
         break
 
     image = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-
-    #sobelx8u = cv2.Sobel(image, cv2.CV_8U, 1, 0, ksize=15)
-    # Output dtype = cv.CV_64F. Then take its absolute and convert to cv.CV_8U
-    #sobelx64f = cv2.Sobel(image, cv2.CV_64F, 1, 0, ksize=5)
-    #abs_sobelx64f = np.absolute(sobelx64f)
-    #sobelx_8u = np.uint8(abs_sobelx64f)
-
-    #sobely64f = cv2.Sobel(image, cv2.CV_64F, 0, 1, ksize=5)
-    #abs_sobely64f = np.absolute(sobely64f)
-    #sobely_8u = np.uint8(abs_sobely64f)
 
     laplacian = cv2.Laplacian(image, cv2.CV_8U)
     laplacian = laplacian.astype(int)
@@ -93,17 +85,19 @@ total_image = total_image.astype(np.uint8)
 
 #print(total_image)
 
-cv2.imshow('img', total_image)
-cv2.waitKey(0)
-cv2.destroyAllWindows()
+if(debug):
+    cv2.imshow('img', total_image)
+    cv2.waitKey(0)
+    cv2.destroyAllWindows()
 
 total_image = cv2.threshold(total_image, 20, 255, cv2.THRESH_BINARY)[1]
 #edges = cv2.Canny(total_image,0,0)
 edges = total_image
 
-cv2.imshow('img', total_image)
-cv2.waitKey(0)
-cv2.destroyAllWindows()
+if(debug):
+    cv2.imshow('img', total_image)
+    cv2.waitKey(0)
+    cv2.destroyAllWindows()
 
 height, width = total_image.shape[:2]
 linesize = int(min(width, height)/8)
@@ -200,6 +194,10 @@ for i in range(3, height - 3):
         maxr = i
 
 
+if(maxce - maxcs < linesize or maxre - maxrs < linesize):
+    print("Did not detect two lines")
+    quit()
+
 cap = cv2.VideoCapture(sys.argv[1])
 ret, image = cap.read()
 
@@ -207,62 +205,64 @@ ret, image = cap.read()
 cv2.line(image, (maxc, maxcs), (maxc, maxce), (0,0,255), 3, cv2.LINE_AA)
 cv2.line(image, (maxrs, maxr), (maxre, maxr), (0,0,255), 3, cv2.LINE_AA)
 
-cv2.imshow('img', image)
-cv2.waitKey(0)
-cv2.destroyAllWindows()
+if(debug):
+    cv2.imshow('img', image)
+    cv2.waitKey(0)
+    cv2.destroyAllWindows()
 
-#lines = cv2.HoughLines(edges,1,np.pi/180,linesize)
-#
-#tolerance = 1
-#index = 0
-#maxlines = 220
-#intervals = 180
-#slotwidth = 360 // intervals
-#ortho_dist = 90 // slotwidth
-#if(90 % slotwidth != 0):
-#    print("Error, bad number of intervals. Needs to be divisible by 90 or something. Not entirely sure, math is hard")
-## stores lines as tuples (x0, y0, dx=b, dy=a) based on angle
-#
-#xs = []
-#ys = []
-#
-#print(lines.shape)
-#
-#if lines is not None:
-#    for i in range(0, len(lines)):
-#        index += 1
-#        if(index > maxlines):
-#            break
-#
-#        rho = lines[i][0][0]
-#        theta = lines[i][0][1]
-#        a = math.cos(theta)
-#        b = math.sin(theta)
-#        x0 = a * rho
-#        y0 = b * rho
-#        pt1 = (int(x0 + 1000*(-b)), int(y0 + 1000*(a)))
-#        pt2 = (int(x0 - 1000*(-b)), int(y0 - 1000*(a)))
-#        #cv2.line(img, pt1, pt2, (0,0,255), 3, cv2.LINE_AA)
-#
-#        angle = math.atan2(a, b) * 180/np.pi
-#        if(angle < 0):
-#            angle += 360
-#
-#
-#
-#        horiz = (angle <= tolerance or angle >= 360 - tolerance or abs(angle - 180) <= tolerance)
-#        vert  = (abs(angle - 90) <= tolerance or abs(angle - 270) <= tolerance)
-#
-#        valid = (horiz and y0 > padding and y0 < height - padding) or (vert and x0 > padding and x0 < width - padding)
-#
-#        if(valid and (horiz or vert)):
-#            cv2.line(total_image, pt1, pt2, (255,255,255), 3, cv2.LINE_AA)
-#
-#        if(valid and horiz):
-#            ys.append(y0)
-#        if(valid and vert):
-#            xs.append(x0)
-#
-#cv2.imshow('img', total_image)
-#cv2.waitKey(0)
-#cv2.destroyAllWindows()
+wig = 2 # wiggle room for lines to overlap
+
+right = False
+left = False
+up = False
+down = False
+
+if(maxrs >= maxc - wig):
+    right = True
+
+if(maxre <= maxc + wig):
+    left = True
+
+if(maxce <= maxr + wig):
+    up = True
+
+if(maxcs >= maxr - wig):
+    down = True
+
+p1 = (maxc, maxr)
+
+p2 = -1
+
+if(up and right):
+    p2 = (width-1, 0)
+
+if(up and left):
+    p2 = (0, 0)
+
+if(down and right):
+    p2 = (width-1, height-1)
+
+if(down and left):
+    p2 = (0, height-1)
+
+if(p2 == -1):
+    print("No obvious box")
+    quit()
+
+xl = min(p1[0], p2[0])
+xh = max(p1[0], p2[0])
+yl = min(p1[1], p2[1])
+yh = max(p1[1], p2[1])
+
+video = VideoFileClip(sys.argv[1])
+boxvideo = video.crop(x1 = xl, y1 = yl, x2 = xh, y2 = yh)
+
+boxvideo.write_videofile("boxvideo.mp4")
+
+def drawrect(frame):
+    cv2.rectangle(frame, (xl, yl), (xh, yh), (0,0,0), -1)
+    return frame
+
+outervideo = video.fl_image(drawrect)
+
+outervideo.write_videofile("outervideo.mp4")
